Add author association to blogs with migration and update related components

+++ database/migrations/2025_02_19_171054_add_author_id_on_blogs_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('blogs', function (Blueprint $table) {
-          $table->unsignedInteger('author_id')->nullable();
+          $table->unsignedBigInteger('author_id')->nullable();
           $table->foreign('author_id')->references('id')->on('users');
         });
     }
